Виджеты разделены по файлам, Рабочая консоль теперь тоже виджет - MultilineEdit
