+++ SimpleLab/GUI/guiFuncs.cpp
