+++ SimpleLab/prefs.h
@@ -3,7 +3,7 @@
 const float BLINK_SPEED = 0.1f;
 const int SCR_W = 480;
 const int SCR_H = 800;
-const float SWIPE_DISTANCE = 60.f;
+const float SWIPE_DISTANCE = 10.f;
 const int MAX_LINES = 10;
 const float TEXT_SCALE_X = 1.f;
 const float TEXT_SCALE_Y = 1.f;
